# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface TrendingHabit {
  id: number;
  question: string;
  meTooPct: number;
  answerCount: number;
  category: string;
  /** @nullable */
  userAnswer: string | null;
}

export interface TrendingResponse {
  mostRelatable: TrendingHabit | null;
  mostSurprising: TrendingHabit | null;
  mostDivisive: TrendingHabit | null;
  newFromCommunity: TrendingHabit[];
}

export interface CategorySummary {
  category: string;
  count: number;
}

export interface HealthStatus {
  status: string;
}

export type HabitTraits = {[key: string]: number};

export interface Habit {
  id: number;
  question: string;
  category: string;
  meTooPct: number;
  traits: HabitTraits;
  source: string;
  status: string;
  answerCount: number;
  reportCount: number;
  flagged: boolean;
}

export interface ReportInput {
  sessionId: string;
  reason: string;
}

export interface ReportResult {
  reported: boolean;
  flagged: boolean;
  reportCount: number;
}

export type FlaggedHabitTopReasonsItem = {
  reason: string;
  count: number;
};

export interface FlaggedHabit {
  id: number;
  question: string;
  category: string;
  source: string;
  status: string;
  reportCount: number;
  flagged: boolean;
  topReasons: FlaggedHabitTopReasonsItem[];
}

export interface SignupInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface AdminAuth {
  pin: string;
}

export interface AdminAuthResponse {
  ok: boolean;
}

export interface AdminHabit {
  id: number;
  question: string;
  category: string;
  source: string;
  status: string;
  answerCount: number;
  meTooPct: number;
  reportCount: number;
  flagged: boolean;
  createdAt: string;
}

export interface FlaggedHabitUpdate {
  /** dismiss | archive */
  action?: string;
}

export interface HabitUpdate {
  question?: string;
  category?: string;
  status?: string;
}

export interface Submission {
  id: number;
  question: string;
  status: string;
  submittedAt: string;
  /** @nullable */
  habitId?: number | null;
  /** @nullable */
  moderationReason?: string | null;
}

export interface MySubmission {
  id: number;
  question: string;
  status: string;
  submittedAt: string;
  /** @nullable */
  habitId?: number | null;
  /** @nullable */
  meTooPct?: number | null;
  answerCount: number;
  reportCount: number;
}

export interface SubmissionInput {
  /**
     * @minLength 10
     * @maxLength 280
     */
  question: string;
  sessionId?: string;
}

export interface SubmissionUpdate {
  status?: string;
  question?: string;
  category?: string;
  /** @nullable */
  moderationReason?: string | null;
}

export interface Answer {
  id: number;
  habitId: number;
  sessionId: string;
  answer: string;
  answeredAt: string;
}

export interface AnswerInput {
  habitId: number;
  sessionId: string;
  answer: string;
}

export type TraitScoreDataScores = {[key: string]: number};

export type TraitScoreDataMaxScores = {[key: string]: number};

export interface TraitScoreData {
  sessionId: string;
  scores: TraitScoreDataScores;
  maxScores: TraitScoreDataMaxScores;
}

export type TraitScoreInputScores = {[key: string]: number};

export type TraitScoreInputMaxScores = {[key: string]: number};

export interface TraitScoreInput {
  scores: TraitScoreInputScores;
  maxScores: TraitScoreInputMaxScores;
}

export type GetTrendingParams = {
sessionId?: string;
};

export type GetHabitsByCategoryParams = {
category: string;
sessionId?: string;
};

export type GetMySubmissionsParams = {
sessionId?: string;
};

export type ListSubmissionsParams = {
status?: string;
};