# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: habits
    description: Habit feed
  - name: answers
    description: User swipe answers
  - name: submissions
    description: User habit submissions
  - name: admin
    description: Admin operations
  - name: traitScores
    description: Per-session trait scores
  - name: auth
    description: Authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /habits:
    get:
      operationId: getHabits
      tags: [habits]
      summary: List active habits with real-time percentages
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"

  /habits/trending:
    get:
      operationId: getTrending
      tags: [habits]
      summary: Trending habits — most relatable, surprising, divisive, and newest
      parameters:
        - in: query
          name: sessionId
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Trending habit groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendingResponse"

  /habits/categories:
    get:
      operationId: getCategories
      tags: [habits]
      summary: List all habit categories with counts
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"

  /habits/by-category:
    get:
      operationId: getHabitsByCategory
      tags: [habits]
      summary: Habits filtered by category with optional session answers
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: true
        - in: query
          name: sessionId
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Habits in category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendingHabit"
        "400":
          description: Missing category

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  /answers/voted:
    get:
      operationId: getVotedHabits
      tags: [answers]
      summary: Get habit IDs the current user has already voted on
      responses:
        "200":
          description: List of voted habit IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        "401":
          description: Not authenticated

  /answers:
    post:
      operationId: recordAnswer
      tags: [answers]
      summary: Record a swipe answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "201":
          description: Answer recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "400":
          description: Invalid input

  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a new habit anonymously
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Invalid input

  /submissions/mine:
    get:
      operationId: getMySubmissions
      tags: [submissions]
      summary: Get current user's submissions with live habit stats
      parameters:
        - in: query
          name: sessionId
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of user's submissions with habit stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MySubmission"
        "400":
          description: Missing sessionId

  /admin/submissions:
    get:
      operationId: listSubmissions
      tags: [admin]
      summary: List all submissions (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"

  /admin/submissions/{id}:
    patch:
      operationId: updateSubmission
      tags: [admin]
      summary: Approve, reject, or edit a submission
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Invalid input
        "404":
          description: Not found

  /habits/{id}/report:
    post:
      operationId: reportHabit
      tags: [habits]
      summary: Report a habit
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Report recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
        "400":
          description: Invalid input
        "404":
          description: Habit not found
        "409":
          description: Already reported

  /admin/flagged-habits:
    get:
      operationId: getFlaggedHabits
      tags: [admin]
      summary: List all flagged habits with report breakdown
      responses:
        "200":
          description: List of flagged habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlaggedHabit"

  /admin/flagged-habits/{id}:
    patch:
      operationId: updateFlaggedHabit
      tags: [admin]
      summary: Dismiss flag or archive a flagged habit
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlaggedHabitUpdate"
      responses:
        "200":
          description: Updated
        "404":
          description: Not found

  /admin/auth:
    post:
      operationId: verifyAdminPin
      tags: [admin]
      summary: Verify admin PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAuth"
      responses:
        "200":
          description: Auth result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"

  /admin/habits:
    get:
      operationId: listAdminHabits
      tags: [admin]
      summary: List all habits with live vote stats (admin)
      responses:
        "200":
          description: All habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminHabit"

  /admin/habits/{id}:
    patch:
      operationId: updateHabit
      tags: [admin]
      summary: Edit a habit (admin)
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
        "404":
          description: Not found
    delete:
      operationId: deleteHabit
      tags: [admin]
      summary: Archive a habit (admin)
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Archived
        "404":
          description: Not found

  /trait-scores/{sessionId}:
    get:
      operationId: getTraitScores
      tags: [traitScores]
      summary: Get trait scores for a session
      parameters:
        - in: path
          name: sessionId
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Trait scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraitScoreData"
        "404":
          description: Not found
    put:
      operationId: upsertTraitScores
      tags: [traitScores]
      summary: Create or replace trait scores for a session
      parameters:
        - in: path
          name: sessionId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TraitScoreInput"
      responses:
        "200":
          description: Upserted trait scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraitScoreData"
        "400":
          description: Invalid input

components:
  schemas:
    TrendingHabit:
      type: object
      required: [id, question, meTooPct, answerCount, category, userAnswer]
      properties:
        id:
          type: integer
        question:
          type: string
        meTooPct:
          type: number
        answerCount:
          type: integer
        category:
          type: string
        userAnswer:
          type: ["string", "null"]

    TrendingResponse:
      type: object
      required: [mostRelatable, mostSurprising, mostDivisive, newFromCommunity]
      properties:
        mostRelatable:
          oneOf:
            - $ref: "#/components/schemas/TrendingHabit"
            - type: "null"
        mostSurprising:
          oneOf:
            - $ref: "#/components/schemas/TrendingHabit"
            - type: "null"
        mostDivisive:
          oneOf:
            - $ref: "#/components/schemas/TrendingHabit"
            - type: "null"
        newFromCommunity:
          type: array
          items:
            $ref: "#/components/schemas/TrendingHabit"

    CategorySummary:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Habit:
      type: object
      required: [id, question, category, meTooPct, traits, source, status, answerCount, reportCount, flagged]
      properties:
        id:
          type: integer
        question:
          type: string
        category:
          type: string
        meTooPct:
          type: number
        traits:
          type: object
          additionalProperties:
            type: number
        source:
          type: string
        status:
          type: string
        answerCount:
          type: integer
        reportCount:
          type: integer
        flagged:
          type: boolean

    ReportInput:
      type: object
      required: [sessionId, reason]
      properties:
        sessionId:
          type: string
        reason:
          type: string

    ReportResult:
      type: object
      required: [reported, flagged, reportCount]
      properties:
        reported:
          type: boolean
        flagged:
          type: boolean
        reportCount:
          type: integer

    FlaggedHabit:
      type: object
      required: [id, question, category, source, status, reportCount, flagged, topReasons]
      properties:
        id:
          type: integer
        question:
          type: string
        category:
          type: string
        source:
          type: string
        status:
          type: string
        reportCount:
          type: integer
        flagged:
          type: boolean
        topReasons:
          type: array
          items:
            type: object
            required: [reason, count]
            properties:
              reason:
                type: string
              count:
                type: integer

    SignupInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthUser:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"

    AdminAuth:
      type: object
      required: [pin]
      properties:
        pin:
          type: string

    AdminAuthResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    AdminHabit:
      type: object
      required: [id, question, category, source, status, answerCount, meTooPct, reportCount, flagged, createdAt]
      properties:
        id:
          type: integer
        question:
          type: string
        category:
          type: string
        source:
          type: string
        status:
          type: string
        answerCount:
          type: integer
        meTooPct:
          type: number
        reportCount:
          type: integer
        flagged:
          type: boolean
        createdAt:
          type: string
          format: date-time

    FlaggedHabitUpdate:
      type: object
      properties:
        action:
          type: string
          description: "dismiss | archive"

    HabitUpdate:
      type: object
      properties:
        question:
          type: string
        category:
          type: string
        status:
          type: string

    Submission:
      type: object
      required: [id, question, status, submittedAt]
      properties:
        id:
          type: integer
        question:
          type: string
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
        habitId:
          type: ["integer", "null"]
        moderationReason:
          type: ["string", "null"]

    MySubmission:
      type: object
      required: [id, question, status, submittedAt, answerCount, reportCount]
      properties:
        id:
          type: integer
        question:
          type: string
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
        habitId:
          type: ["integer", "null"]
        meTooPct:
          type: ["number", "null"]
        answerCount:
          type: integer
        reportCount:
          type: integer

    SubmissionInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 10
          maxLength: 280
        sessionId:
          type: string

    SubmissionUpdate:
      type: object
      properties:
        status:
          type: string
        question:
          type: string
        category:
          type: string
        moderationReason:
          type: ["string", "null"]

    Answer:
      type: object
      required: [id, habitId, sessionId, answer, answeredAt]
      properties:
        id:
          type: integer
        habitId:
          type: integer
        sessionId:
          type: string
        answer:
          type: string
        answeredAt:
          type: string
          format: date-time

    AnswerInput:
      type: object
      required: [habitId, sessionId, answer]
      properties:
        habitId:
          type: integer
        sessionId:
          type: string
        answer:
          type: string

    TraitScoreData:
      type: object
      required: [sessionId, scores, maxScores]
      properties:
        sessionId:
          type: string
        scores:
          type: object
          additionalProperties:
            type: number
        maxScores:
          type: object
          additionalProperties:
            type: number

    TraitScoreInput:
      type: object
      required: [scores, maxScores]
      properties:
        scores:
          type: object
          additionalProperties:
            type: number
        maxScores:
          type: object
          additionalProperties:
            type: number